Translate another part of Segment_1d12 to managed code

+++ src/CivGame/CivGameGlobals.cs
@@ -67,19 +67,19 @@ namespace OpenCiv1
 		public ushort Var_2f9e_Unknown = 0xffff;
 
 		// 0x652e - after this offset the default values are set to 0
-		public ushort Var_6b64 = 0;
-		public ushort Var_6b92 = 0;
-		public ushort Var_6c98 = 0;
-		public ushort Var_6c9a = 0;
+		public int Var_6b64 = 0;
+		public int Var_6b92 = 0;
+		public int Var_6c98 = 0;
+		public int Var_6c9a = 0;
 		public int[] Var_70da_Arr = new int[4];
 		public int Var_70e2 = 0;
 		public int Var_70e4 = 0;
 		public int Var_70e6 = 0;
-		public ushort Var_8078 = 0;
-		public ushort Var_b1e8 = 0;
-		public ushort Var_b882 = 0;
+		public int Var_8078 = 0;
+		public int Var_b1e8 = 0;
+		public int Var_b882 = 0;
 		public ushort Var_b2ba = 0;
-		public ushort Var_d2de = 0;
+		public int Var_d2de = 0;
 		public int Var_d2e0 = 0;
 		public int Var_d2f6 = 0;
 		public short Var_d4cc_XPos = 0;
@@ -93,7 +93,7 @@ namespace OpenCiv1
 		public int Var_deb8 = 0;
 		public int Var_e17a = 0;
 		public int Var_e3c2 = 0;
-		public ushort Var_e3c6 = 0;
+		public int Var_e3c6 = 0;
 		public int Var_e8b8 = 0;
 		#endregion
 	}
